Refactor genotype_main out of main.
